Updated rapidjson. Improved message length by 30%.

+++ Sonic/Sonic.Framework/Vector2.cpp
@@ -3,7 +3,7 @@
 Vector2::Vector2() {
 }
 
-Vector2::Vector2(float x, float y) {
+Vector2::Vector2(double x, double y) {
 	this->x = x;
 	this->y = y;
 }
