+++ Sonic/Sonic.Framework/Vector2.h
@@ -4,10 +4,10 @@
 class Vector2 {
 public:
 	Vector2();
-	Vector2(float x, float y);
+	Vector2(double x, double y);
 
-	float x;
-	float y;
+	double x;
+	double y;
 };
 
 #endif // !VECTOR2_H
